Extensive rewrite of abstract, removal of cruft, fixes #2, #5, #6.
The abstract has been rewritten extensively in order to reflect the main
syntactic properties of annotation as well as why this syntax has been
selected as the most likely candidate in the introductory section.

The initial and obviously boring examples of pattern matching through
parameter patterns have been removed since the document can offer a
much better alternative as to what vertical and horizontal expansion
signify in a very succint manner within the abstract. This however is
readily evident mostly to experts (e.g. boost.preprocessor users).

The section on introduction partially addresses also issue #7, while
within 2-3 paragraphs it is attempted to clarify that parameter pack
processing features turn typelists into first-class citizens of within
a template parameter list.

Examples have been added for getting quickly acquainted with the full
form of the annotation syntax (asides ternary annotation, which is
NOT what you think it is).

References to WG21 papers related to the need of pack processing features
at the language level have been added. The subtle reference to monads
is made through the citation of work by Sinkovic and Porkolab. Concepts
are being hinted upon as being boosted by such semantics while not
able to supplant pack processing features themselves. Ternary pack
annotation is introduced and targeted for the generative aspects of
pack processing but will be discussed later on. These last two (concepts
and ternaries) are actively processed but the other sections have to
be cleaned up or it is a big mess to tell the whole story.