Introducing syntax tables, making EWG30 more evident through them.
The most glaring flaw of my initial draft was to overestimate the zeal with
which C++ standard committee member would reason about annotation. A new
section on annotation is introduced that deals with just the rules and uses
two tables for showing various combinations of integral constant expressions
with annotation. This will supplant "Pack annotation semantics" which may
be heavy and repetitive to read.

Most importantly, EWG issue 30 is now readily evident as directly addressed.
I have used the term "access annotator" for the individual accessor, which
was my old term. The "access annotator" is a degenerate form of the "interval"
annotator, but lacking any connection to the triple-dot operator used with
packs as we know them today. It offers access by index.