Removing some more cruft, priming for "formal" addendum.
No conceptual changes right now, just some generic rearrangements. It
should be enough for changing the "formalness" everybody dislikes but
is important to reason through into a better form in the addendum.
Deployment scenarios should be moved prior to that and immediately after
annotation rules and their syntax table examples.