Expansion enhancements, 2-expression access annotation and triple-dot.
This helps go straight into "ternary" for which I still have to find a
better term. The interesting interplay with triple-dot and the interval or
access annotator is quite interesting because it allows full expression
semantics to be terse and precise in their contextual meaning. More work
is needed to clarify and shorten this.