Removed a pointless and dangerous shorthand
There is no need and point for the access annotator to be followed by
triple-dot when we want to expand from a given index to the end
of the pack since this can be done through T{N,sizeof...(T)}. Such
a triple-dot addition would make things ambiguous with non-type
parameters for no good reason. Indeed, commit bdb29d2 did not have this
and neither should it exist.