Changing the names of the annotators.
The following terms substitute the previously used ones for annotators:

   1) "size  annotator" substitutes: "interval annotator"
   2) "tuple annotator" substitutes: "pattern annotator"
   3) "range annotator" substitutes: "access annotator"

The reason for this change is that it becomes easier to explain what the
annotators do because the interval annotator made a size-based match over
and inverval of sizes, with the access annotator expanding to the ordered
sequence of types within specified index bounds. Both should thus be
considered as 2-expression based degenerating into 1-expression based
annotation when we degenerate from intervals of sizes and types to
a specific size or type. This can really shorten the text for what this
draft actually discusses.

The tuple annotator practically has the effect of providing a check on
whether a pack is a tuple of pack whose size is sizeof...(T)/K where K
is the integral constant expression of the annotator previously known
as "pattern annotator".

Other changes included removing some sections that need no longer be in
such form, especially with the "formal" part which despite correct, was
quite boring for "learn by example people". It will make a comeback
later on in the addendum.

Fixes #12, fixes #9, fixes #7 due to issue deprecation and renaming.