More interdependent and formal analysis derived annotator definitions.
Removed a the "Getting quickly acquainted with the syntax" section since
we no longer really need it; the syntax table and the rules should be
adequate prior to immediately jumping to a use tutorial. The order
of presentation of the annotators has been changed, with the tuple
annotator being last. Their definitions are more interdependent now,
as is derived from the formal analysis parts that are currently
withheld from the document. I have to add a better tutorial and after
that finish with the formal parts in the appendix; the result will
be a very shorter document but quite usable in proving the point that
was made.