In relation to issue #13, \"Multiplicity\"
The semantics of which kinds of annotation are valid to allow multiple
annotated parameter packs to coexist in the same parameter list have
to be properly specified; this is easy for 1-expression size annotation,
but quite more delicate for 2-expression size annotation. This branch
addresses issue #13 in the atpp bugtracker.