Fixing a semantic clash between syntax table and annotation text.
The syntax table used for illustrating annotation had a copy/paste error
during document editing that did not respect changes made to the atpp
draft already in commit e7dd3706f2868c983e86898cea8f4f4e263179c1 and
before, including the really old bdb29d245984bd8cca2213d4d19ab0cb6257c573.

Specifically, it is obvious that the "tuple" annotator does not reflect
a K-tuple of sizeof...(T)/K tuples as the syntax table implies, but
any tuple consisting of packs of the first sizeof...(T)/K parameters
repeated identically. This allows for the following obvious example:

    template<typename... T{}[sizeof...(T)]>
    void example(T...) {}

Given that sizeof...(T)/sizeof...(T) == 1 the entire pack T must
be comprised of sizeof...(T) parameters abstracting the "1"-tuple
of the first "1" parameters of the parameter pack T. Therefore
all the template type parameters must be abstracting the same type
that is the "head" of the pack. This is possible because in the
context of annotation semantics, pack identifiers can participate
in integral constant expressions involving them within annotation;
they have already been specified after all. This behavior was
already implied in bdb29d245984bd8cca2213d4d19ab0cb6257c573 and
former, so the K-tuple to any-tuple is an obvious editing correction,
given that the semantics of annotation as explained outside the
syntax table are already expressing exactly that.

Also, this should be added in the deployment examples section.