Changing order of appearance, range annotator wording.
It makes more sense to have the first examples be of specific to
just the size annotator in the syntax table instead of how they
were. Also, some clarifications on how range annotation offers
individual and by index range access. Another wording detail is
needed in order to make this completely clear in use context, after
the examples that are being prepared.