Fixes #3, fixes #8
Because of the introduction of the equivalence table, we should not rely
on the "math" part that everybody found tedious in Rapperswil; that can
wait. Given that we have initial examples of fmap, foldl, foldr in a
library using annotation, explaining this with symbolic notation is quite
superfluous, so I am removing the math section completely at this point.

The introduction of the fmap and the folds should eventually be ironed in
sequence with the previous examples.